+++ vitest.config.js
@@ -1,0 +1,5 @@
+export default {
+  test: {
+    testTimeout: 30_000
+  }
+};